Damage and rate rate corretly scalable
support modifier changed from public to private. This solved the fire rate problem.
Damage of the turret is now also set onto the bullet when instantiating with bullet.damage = damage. This effect is turret specific and does not affect other turrets.

+++ Assets/Scripts/Bullet.cs
@@ -6,7 +6,7 @@ public class Bullet : MonoBehaviour
     private Transform target;
 
     public float speed = 70f;
-    public int damage = 50;
+    public float damage = 50;
     public float explosionRadius = 0f;
     public GameObject impactEffect;
     [HideInInspector]
@@ -20,6 +20,7 @@ public class Bullet : MonoBehaviour
 
     void Start()
     {
+
         if (target != null)
         {
             dir = target.position - transform.position;
